The best and the shittiest SASL script for mIRC, so far.
Newest (2017) mIRC script for SASL support. Works with all versions of mIRC from 6.10 through 7.47+. Utilizes mIRC's internal server settings and UI for password management. Sends through SASL auth blindly to the server, but works. Goal of this script is to implement SASL with the least complexity, easiest to read, and greatest client / server compatibility. It does not violate IRCv3 protocol, per se, as the script doesn't do anything that it MUST NOT do, it just forgivingly ignores things it SHOULD or MAY do.

Consideration is requested for publicly visible publishing. The word SHITTY should not be a deterrent.