List all ports together for clarity.
I think it makes sense to list all the ports together, so it's immediately
obvious which ones are plain text and which are SSL, and for easy copy & paste.

I don't really like the fact it leaves the SSL ones duplicated in the SSL
section right below, though - if that irks others, maybe it could be re-worded
to e.g. "offers SSL access on the SSL ports noted above" or something.