Replace mIRC SASL page with built-in SASL info
mIRC added built-in SASL support in version 7.48. Replace old instructions with the built-in instructions, because the old ones 1) allegedly don't work on current mIRC and 2) just encourage people to use out-of-date versions, which is really not optimal.

These instructions were written and tested by me, using mIRC 7.51 (current as of right now).