SASL:irssi: add `cd` line when downloading script (#470)
Instructions were missing a `cd` line to actually change directory to ~/.irssi/scripts/autorun, which meant the `../cap_sasl.pl` in the last line wouldn't point to the right place. Since having the symlink be relative is useful (in case ~/.irssi/ gets moved later or something), let's add a `cd` and update everything to be relative, rather than making the path in `ln` absolute.